+++ lib/aerospike/command/command.rb
@@ -58,7 +58,8 @@ module Aerospike
   INFO2_DURABLE_DELETE = Integer(1 << 4)
   # Create only. Fail if record already exists.
   INFO2_CREATE_ONLY = Integer(1 << 5)
-
+  # Treat as long query, but relax read consistency.
+  INFO2_RELAX_AP_LONG_QUERY = (1 << 6)
   # Return a result for every operation.
   INFO2_RESPOND_ALL_OPS = Integer(1 << 7)
 
@@ -195,7 +196,7 @@ module Aerospike
       field_count += 1 if exp_size > 0
 
       size_buffer
-      write_header_read(policy, INFO1_READ | INFO1_GET_ALL, 0, field_count, 0)
+      write_header_read(policy, INFO1_READ | INFO1_GET_ALL, 0, 0, field_count, 0)
       write_key(key)
       write_filter_exp(@policy.filter_exp, exp_size)
       end_cmd
@@ -220,7 +221,7 @@ module Aerospike
           attr |= INFO1_GET_ALL
         end
 
-        write_header_read(policy, attr, 0, field_count, bin_names.length)
+        write_header_read(policy, attr, 0, 0, field_count, bin_names.length)
         write_key(key)
         write_filter_exp(@policy.filter_exp, exp_size)
 
@@ -377,7 +378,7 @@ module Aerospike
         operation_count = bin_names.length
       end
 
-      write_header_read(policy, read_attr, info_attr, field_count, operation_count)
+      write_header_read(policy, read_attr, 0, info_attr, field_count, operation_count)
 
       if namespace
         write_field_string(namespace, Aerospike::FieldType::NAMESPACE)
@@ -591,10 +592,16 @@ module Aerospike
         write_header_write(policy, INFO2_WRITE, field_count, operation_count)
       else
         read_attr = INFO1_READ
+        write_attr = 0
+
         read_attr |= INFO1_NOBINDATA unless policy.include_bin_data
-        read_attr |= INFO1_SHORT_QUERY if policy.short_query
+        if policy.short_query || policy.expected_duration == QueryDuration::SHORT
+          read_attr |= INFO1_SHORT_QUERY
+        elsif policy.expected_duration == QueryDuration::LONG_RELAX_AP
+          write_attr |= INFO2_RELAX_AP_LONG_QUERY
+        end
         info_attr = INFO3_PARTITION_DONE if is_new
-        write_header_read(policy, read_attr, info_attr, field_count, operation_count)
+        write_header_read(policy, read_attr, write_attr, info_attr, field_count, operation_count)
       end
 
 
@@ -960,13 +967,13 @@ module Aerospike
       @data_offset = MSG_TOTAL_HEADER_SIZE
     end
 
-    def write_header_read(policy, read_attr, info_attr, field_count, operation_count)
+    def write_header_read(policy, read_attr, write_attr, info_attr, field_count, operation_count)
       read_attr |= INFO1_COMPRESS_RESPONSE if policy.use_compression
       #TODO: Add SC Mode
 
       @data_buffer.write_byte(MSG_REMAINING_HEADER_SIZE, 8) # Message header.length.
       @data_buffer.write_byte(read_attr, 9)
-      @data_buffer.write_byte(0, 10)
+      @data_buffer.write_byte(write_attr, 10)
       @data_buffer.write_byte(info_attr, 11)
 
       (12...18).each { |i| @data_buffer.write_byte(0, i) }